REST implemented

+++ src/main/java/com/example/spring/rest/entities/Category.java
@@ -17,7 +17,7 @@ public class Category {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Byte id;
 
     private String  name;
